split CFFCharset to two classes
