+++ fontbox/src/main/java/org/apache/fontbox/cff/CFFCharset.java
@@ -16,46 +16,19 @@
  */
 package org.apache.fontbox.cff;
 
-import java.util.HashMap;
-import java.util.Map;
-
 /**
  * A CFF charset. A charset is an array of SIDs/CIDs for all glyphs in the font.
- *
- * todo: split this into two? CFFCharsetType1 and CFFCharsetCID ?
  *
  * @author John Hewson
  */
 public abstract class CFFCharset
 {
-    private final boolean isCIDFont;
-    private final Map<Integer, Integer> sidOrCidToGid = new HashMap<>(250);
-    private final Map<Integer, Integer> gidToSid = new HashMap<>(250);
-    private final Map<String, Integer> nameToSid = new HashMap<>(250);
-
-    // inverse
-    private final Map<Integer, Integer> gidToCid = new HashMap<>();
-    private final Map<Integer, String> gidToName = new HashMap<>(250);
-
-    /**
-     * Package-private constructor for use by subclasses.
-     *
-     * @param isCIDFont true if the parent font is a CIDFont
-     */
-    CFFCharset(boolean isCIDFont)
-    {
-        this.isCIDFont = isCIDFont;
-    }
-
     /**
      * Indicates if the charset belongs to a CID font.
      * 
      * @return true for CID fonts
      */
-    public boolean isCIDFont()
-    {
-        return isCIDFont;
-    }
+    public abstract boolean isCIDFont();
     
     /**
      * Adds a new GID/SID/name combination to the charset.
@@ -63,17 +36,7 @@ public abstract class CFFCharset
      * @param gid GID
      * @param sid SID
      */
-    public void addSID(int gid, int sid, String name)
-    {
-        if (isCIDFont)
-        {
-            throw new IllegalStateException("Not a Type 1-equivalent font");
-        }
-        sidOrCidToGid.put(sid, gid);
-        gidToSid.put(gid, sid);
-        nameToSid.put(name, sid);
-        gidToName.put(gid, name);
-    }
+    public abstract void addSID(int gid, int sid, String name);
 
     /**
      * Adds a new GID/CID combination to the charset.
@@ -81,15 +44,7 @@ public abstract class CFFCharset
      * @param gid GID
      * @param cid CID
      */
-    public void addCID(int gid, int cid)
-    {
-        if (!isCIDFont)
-        {
-            throw new IllegalStateException("Not a CIDFont");
-        }
-        sidOrCidToGid.put(cid, gid);
-        gidToCid.put(gid, cid);
-    }
+    public abstract void addCID(int gid, int cid);
 
     /**
      * Returns the SID for a given GID. SIDs are internal to the font and are not public.
@@ -97,19 +52,7 @@ public abstract class CFFCharset
      * @param sid SID
      * @return GID
      */
-    int getSIDForGID(int sid)
-    {
-        if (isCIDFont)
-        {
-            throw new IllegalStateException("Not a Type 1-equivalent font");
-        }
-        Integer gid = gidToSid.get(sid);
-        if (gid == null)
-        {
-            return 0;
-        }
-        return gid;
-    }
+    abstract int getSIDForGID(int sid);
 
     /**
      * Returns the GID for the given SID. SIDs are internal to the font and are not public.
@@ -117,19 +60,7 @@ public abstract class CFFCharset
      * @param sid SID
      * @return GID
      */
-    int getGIDForSID(int sid)
-    {
-        if (isCIDFont)
-        {
-            throw new IllegalStateException("Not a Type 1-equivalent font");
-        }
-        Integer gid = sidOrCidToGid.get(sid);
-        if (gid == null)
-        {
-            return 0;
-        }
-        return gid;
-    }
+    abstract int getGIDForSID(int sid);
 
     /**
      * Returns the GID for a given CID. Returns 0 if the CID is missing.
@@ -137,19 +68,7 @@ public abstract class CFFCharset
      * @param cid CID
      * @return GID
      */
-    public int getGIDForCID(int cid)
-    {
-        if (!isCIDFont)
-        {
-            throw new IllegalStateException("Not a CIDFont");
-        }
-        Integer gid = sidOrCidToGid.get(cid);
-        if (gid == null)
-        {
-            return 0;
-        }
-        return gid;
-    }
+    public abstract int getGIDForCID(int cid);
 
     /**
      * Returns the SID for a given PostScript name, you would think this is not needed,
@@ -158,19 +77,7 @@ public abstract class CFFCharset
      * @param name PostScript glyph name
      * @return SID
      */
-    int getSID(String name)
-    {
-        if (isCIDFont)
-        {
-            throw new IllegalStateException("Not a Type 1-equivalent font");
-        }
-        Integer sid = nameToSid.get(name);
-        if (sid == null)
-        {
-            return 0;
-        }
-        return sid;
-    }
+    abstract int getSID(String name);
 
     /**
      * Returns the PostScript glyph name for the given GID.
@@ -178,14 +85,7 @@ public abstract class CFFCharset
      * @param gid GID
      * @return PostScript glyph name
      */
-    public String getNameForGID(int gid)
-    {
-        if (isCIDFont)
-        {
-            throw new IllegalStateException("Not a Type 1-equivalent font");
-        }
-        return gidToName.get(gid);
-    }
+    public abstract String getNameForGID(int gid);
 
     /**
      * Returns the CID for the given GID.
@@ -193,18 +93,5 @@ public abstract class CFFCharset
      * @param gid GID
      * @return CID
      */
-    public int getCIDForGID(int gid)
-    {
-        if (!isCIDFont)
-        {
-            throw new IllegalStateException("Not a CIDFont");
-        }
-
-        Integer cid = gidToCid.get(gid);
-        if (cid != null)
-        {
-            return cid;
-        }
-        return 0;
-    }
+    public abstract int getCIDForGID(int gid);
 }
